Add files via upload
This project focuses on analyzing the unemployment rate, specifically its sharp increase during the Covid-19 pandemic. The unemployment rate is defined as the percentage of the labor force that is unemployed but actively seeking employment. This analysis aims to provide insights into how the pandemic has impacted unemployment and to identify patterns and trends over time.